Cache portal data every time portal details are rendered.
This is use when exploring and it comes across a portal with just a
placeholder.  For now, we know these are portals that were moved but still
have a link pointing to the old location.

Since we now trigger a rendering each time we encounter just such a
placeholder, this is a rather convenient way of gathering the data.

The user does still have to restart the exploration, but it is a LOT better
than having to scroll around, resize, and other fun things to accomplish this
same thing manually, and which may not always work.